Moved get_outlier_scores and get_ood_scores to outlier.py
